Same fix for editor and spectator.

+++ src/wui/interactive_player.cc
@@ -309,8 +309,8 @@ void InteractivePlayer::draw_map_view(MapView* given_map_view, RenderTarget* dst
 
 		const float scale = 1.f / given_map_view->view().zoom;
 		const auto blit_overlay = [dst, f, scale](const Image* pic, const Vector2i& hotspot) {
-			const Recti pixel_perfect_rect = Rectf(f->rendertarget_pixel - hotspot.cast<float>() * scale,
-														pic->width() * scale, pic->height() * scale).cast<int>();
+			const Recti pixel_perfect_rect = Recti(f->rendertarget_pixel.cast<int>() - hotspot * scale,
+														pic->width() * scale, pic->height() * scale);
 			dst->blitrect_scale(pixel_perfect_rect.cast<float>(),
 			                    pic, Recti(0, 0, pic->width(), pic->height()), 1.f,
 			                    BlendMode::UseAlpha);
